Allow us to override the branch
When we run ci, we checkout the branch and merge in master. This leaves us in a detached
state. This means that we aren't really on a branch anymore. On deploy, we need to set the
branch to master so that coveralls knows how to compare coverages. This allows us to
override the branch.