php-coveralls: Allow overriding the commit sha
In ci, we merge in master which puts us on a different commit sha which we just
throw away afterwords. If we just do `git log -1`, we get the commit information
of the throwaway merge commit.

Instead, let us override this with the 'correct' information of the HEAD of the 
actual branch.